Rename testPrograms to testAssets Strings should be included which are not programs therefore the generalization.
