+++ src/util/prolog/prolog.test.ts
@@ -1,80 +1,80 @@
 import {importFile} from "./prologSession";
-import * as programs from "./testPrograms";
+import * as assets from "./testAssets";
 
 // Tau-Prolog always uses "false." to signify the end of the output stream. This behavior deviates from
 // SWI Prolog's behavior sometimes, but it generally more consistent.
 
 describe("Tau-Prolog Behavior", () => {
 	test("Evaluates Facts Correctly (True Queries)", async () => {
-		const session = programs.evenNumbersFacts;
+		const session = assets.evenNumbersFacts;
 		const results = (await session.executeQuery("even(2).")).getRawResults();
 
 		expect(results).toStrictEqual(["true ;", "false."]);
 	});
 
 	test("Evaluates Facts Correctly (False Queries)", async () => {
-		const session = programs.evenNumbersFacts;
+		const session = assets.evenNumbersFacts;
 		const results = (await session.executeQuery("even(3).")).getRawResults();
 
 		expect(results).toStrictEqual(["false."]);
 	});
 
 	test("Evaluates Equalities Correctly (True Queries)", async () => {
-		const session = programs.emptyCode;
+		const session = assets.emptyCode;
 		const results = (await session.executeQuery("2 = 2.")).getRawResults();
 
 		expect(results).toStrictEqual(["true ;", "false."]);
 	});
 
 	test("Evaluates Equalities Correctly (False Queries)", async () => {
-		const session = programs.emptyCode;
+		const session = assets.emptyCode;
 		const results = (await session.executeQuery("2 = 3.")).getRawResults();
 
 		expect(results).toStrictEqual(["false."]);
 	});
 
 	test("Evaluates Transitiv Equalities Correctly (False Query)", async () => {
-		const session = programs.emptyCode;
+		const session = assets.emptyCode;
 		const results = (await session.executeQuery("X = 2, X = 3.")).getRawResults();
 
 		expect(results).toStrictEqual(["false."]);
 	});
 
 	test("Does Not Resolve Operators with =", async () => {
-		const session = programs.emptyCode;
+		const session = assets.emptyCode;
 		const results = (await session.executeQuery("X = 2 + 3.")).getRawResults();
 
 		expect(results).toStrictEqual(["X = '+'(2, 3) ;", "false."]);
 	});
 
 	test("Resolves Operators with the 'is'", async () => {
-		const session = programs.emptyCode;
+		const session = assets.emptyCode;
 		const results = (await session.executeQuery("X is 2 + 3.")).getRawResults();
 
 		expect(results).toStrictEqual(["X = 5 ;", "false."]);
 	});
 
 	test("Assigns Values To One Variable Correctly", async () => {
-		const session = programs.evenNumbersFacts;
+		const session = assets.evenNumbersFacts;
 		const results = (await session.executeQuery("even(X).")).getRawResults();
 
 		expect(results).toStrictEqual(["X = 2 ;", "X = 4 ;", "X = 6 ;", "X = 8 ;", "false."]);
 	});
 
 	test("Assigns Values To Two Variables Correctly", async () => {
-		const session = programs.likingFacts;
+		const session = assets.likingFacts;
 		const results = (await session.executeQuery("likes(X, Y).")).getRawResults();
 		expect(results).toStrictEqual(["X = lisa, Y = bob ;", "X = frank, Y = lisa ;", "false."]);
 	});
 
 	test("Evaluates Expression with the Cut-Operator Correctly", async () => {
-		const session = programs.max;
+		const session = assets.max;
 		const results = (await session.executeQuery("max(2, 5, X).")).getRawResults();
 		expect(results).toStrictEqual(["X = 5 ;", "false."]);
 	});
 
 	test("The write/1 function returns true", async () => {
-		const session = programs.emptyCode;
+		const session = assets.emptyCode;
 		const results = (await session.executeQuery("write('hello').")).getRawResults();
 
 		expect(results).toStrictEqual(["true ;", "false."]);
@@ -83,7 +83,7 @@ describe("Tau-Prolog Behavior", () => {
 
 describe("PrologResult.getResults()", () => {
 	test('The evenNumberFacts Program returns the correct Map for "even(x)."', async () => {
-		const session = programs.evenNumbersFacts;
+		const session = assets.evenNumbersFacts;
 		const results = (await session.executeQuery("even(X).")).getResults();
 
 		const expectedResults =
@@ -95,7 +95,7 @@ describe("PrologResult.getResults()", () => {
 	});
 
 	test('The family Program returns the correct Map for "grandmother(X, son)."', async () => {
-		const session = programs.family;
+		const session = assets.family;
 		const results = (await session.executeQuery("grandmother(X, son).")).getResults();
 
 		const expectedResults =
@@ -109,7 +109,7 @@ describe("PrologResult.getResults()", () => {
 
 describe("PrologResult.getResultArray()", () => {
 	test('The evenNumberFacts Program returns [["2"], ["4"],  ["6"], ["8"], ["false"]] for "even(x)."', async () => {
-		const session = programs.evenNumbersFacts;
+		const session = assets.evenNumbersFacts;
 		const results = (await session.executeQuery("even(X).")).getResultArray();
 
 		expect(results).toStrictEqual([["2"], ["4"], ["6"], ["8"], ["false"]]);
@@ -117,7 +117,7 @@ describe("PrologResult.getResultArray()", () => {
 
 	test('The family Program returns [["grandmotherm"], ["grandmotherf"], ["false"]] for "grandmother(X, son)."',
 		async () => {
-			const session = programs.family;
+			const session = assets.family;
 			const results = (await session.executeQuery("grandmother(X, son).")).getResultArray();
 
 			expect(results).toStrictEqual([["grandmotherm"], ["grandmotherf"], ["false"]]);
